requirements.txt
pytest==7.2.2
selenium==4.8.2
webdriver-manager==4.0.0
allure-pytest=2.13.2